WIP: More refactoring of the importers, flattening class hierarchy and making most data-munging methods static (for now.)
