+++ ehri-io/src/main/java/eu/ehri/project/importers/csv/CsvAuthoritativeItemImporter.java
@@ -30,7 +30,7 @@ import eu.ehri.project.acl.SystemScope;
 import eu.ehri.project.definitions.Ontology;
 import eu.ehri.project.exceptions.ValidationError;
 import eu.ehri.project.importers.ImportLog;
-import eu.ehri.project.importers.base.MapImporter;
+import eu.ehri.project.importers.base.AbstractImporter;
 import eu.ehri.project.importers.util.Helpers;
 import eu.ehri.project.models.EntityClass;
 import eu.ehri.project.models.base.Accessible;
@@ -52,7 +52,7 @@ import java.util.Map;
  * Importer of authoritative items such as historical agents and concepts, loaded
  * from a CSV file.
  */
-public class CsvAuthoritativeItemImporter extends MapImporter {
+public class CsvAuthoritativeItemImporter extends AbstractImporter<Map<String, Object>> {
 
     private static final Logger logger = LoggerFactory.getLogger(CsvAuthoritativeItemImporter.class);
 
@@ -116,29 +116,5 @@ public class CsvAuthoritativeItemImporter extends MapImporter {
             Helpers.putPropertyInGraph(item, Ontology.LANGUAGE_OF_DESCRIPTION, "en");
         }
         return item;
-    }
-
-    @Override
-    public List<Map<String, Object>> extractDates(Map<String, Object> itemData) {
-
-        List<Map<String, Object>> l = Lists.newArrayList();
-        Map<String, Object> items = Maps.newHashMap();
-
-        String end = (String) itemData.get("DateofdeathYYYY-MM-DD");
-        String start = (String) itemData.get("DateofbirthYYYY-MM-DD");
-        if (start != null && start.endsWith("00-00")) {
-            start = start.substring(0, 4);
-        }
-        if (end != null && end.endsWith("00-00")) {
-            end = end.substring(0, 4);
-        }
-        if (end != null || start != null) {
-            if (start != null)
-                items.put(Ontology.DATE_PERIOD_START_DATE, start);
-            if (end != null)
-                items.put(Ontology.DATE_PERIOD_END_DATE, end);
-            l.add(items);
-        }
-        return l;
     }
 }
