+++ ehri-io/src/main/java/eu/ehri/project/importers/ead/UshmmHandler.java
@@ -23,6 +23,7 @@ package eu.ehri.project.importers.ead;
 
 import eu.ehri.project.importers.base.AbstractImporter;
 import eu.ehri.project.importers.properties.XmlImportProperties;
+import eu.ehri.project.importers.util.Helpers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,15 +48,15 @@ public class UshmmHandler extends EadHandler {
     @Override
     protected void extractIdentifier(Map<String, Object> currentGraph) {
         //not all units have ids, and some have multiple, find the "irn"
-        if (currentGraph.containsKey(OBJECT_IDENTIFIER)) {
-            if (currentGraph.get(OBJECT_IDENTIFIER) instanceof List) {
-                logger.debug("class of identifier: " + currentGraph.get(OBJECT_IDENTIFIER).getClass());
-                List<String> identifiers = (List<String>) currentGraph.get(OBJECT_IDENTIFIER);
+        if (currentGraph.containsKey(Helpers.OBJECT_IDENTIFIER)) {
+            if (currentGraph.get(Helpers.OBJECT_IDENTIFIER) instanceof List) {
+                logger.debug("class of identifier: " + currentGraph.get(Helpers.OBJECT_IDENTIFIER).getClass());
+                List<String> identifiers = (List<String>) currentGraph.get(Helpers.OBJECT_IDENTIFIER);
                 List<String> identifierType = (List<String>) currentGraph.get("objectIdentifierType");
                 for (int i = 0; i < identifiers.size(); i++) {
                     if (identifierType.get(i).equals("irn")) {
                         logger.debug("found official id: " + identifiers.get(i));
-                        currentGraph.put(OBJECT_IDENTIFIER, identifiers.get(i));
+                        currentGraph.put(Helpers.OBJECT_IDENTIFIER, identifiers.get(i));
                     } else {
                         logger.debug("found other form of identifier: " + identifiers.get(i));
                         addOtherIdentifier(currentGraph, identifiers.get(i));
@@ -66,7 +67,7 @@ public class UshmmHandler extends EadHandler {
             }
         } else {
             logger.error("no unitid found, setting {}", ++count);
-            currentGraph.put(OBJECT_IDENTIFIER, "ushmmID" + count);
+            currentGraph.put(Helpers.OBJECT_IDENTIFIER, "ushmmID" + count);
         }
     }
 }
