+++ ehri-io/src/main/java/eu/ehri/project/importers/eag/EagImporter.java
@@ -22,9 +22,10 @@ package eu.ehri.project.importers.eag;
 import com.tinkerpop.frames.FramedGraph;
 import eu.ehri.project.definitions.Ontology;
 import eu.ehri.project.exceptions.ValidationError;
-import eu.ehri.project.importers.base.SaxXmlImporter;
+import eu.ehri.project.importers.base.AbstractImporter;
 import eu.ehri.project.importers.ImportLog;
 import eu.ehri.project.importers.eac.EacImporter;
+import eu.ehri.project.importers.util.Helpers;
 import eu.ehri.project.models.Country;
 import eu.ehri.project.models.EntityClass;
 import eu.ehri.project.models.Repository;
@@ -45,7 +46,7 @@ import java.util.regex.Pattern;
 /**
  * Importer of EAG-based descriptions.
  */
-public class EagImporter extends SaxXmlImporter {
+public class EagImporter extends AbstractImporter<Map<String, Object>> {
 
     private static final Logger logger = LoggerFactory.getLogger(EacImporter.class);
     private final Pattern priorityPattern = Pattern.compile("Priority: (-?\\d+)");
@@ -69,9 +70,9 @@ public class EagImporter extends SaxXmlImporter {
         return importItem(itemData);
     }
 
-    @Override
     public Map<String, Object> extractUnit(Map<String, Object> itemData) throws ValidationError {
-        Map<String, Object> data = super.extractUnit(itemData);
+        Map<String, Object> data = Helpers.extractIdentifiers(itemData);
+        data.put("typeOfEntity", itemData.get("typeOfEntity"));
         // MB: Hack hack hack - extract EHRI-specific 'priority' field out of the
         // pattern "Priority: <digit>" in the maintenanceNotes field.
         Object notes = itemData.get(MAINTENANCE_NOTES);
@@ -99,27 +100,27 @@ public class EagImporter extends SaxXmlImporter {
 
         BundleManager persister = new BundleManager(framedGraph, permissionScope.idPath());
 
-        Map<String, Object> descmap = extractUnitDescription(itemData, EntityClass.REPOSITORY_DESCRIPTION);
+        Map<String, Object> descmap = Helpers.extractUnitDescription(itemData, EntityClass.REPOSITORY_DESCRIPTION);
         descmap.put(Ontology.IDENTIFIER_KEY, descmap.get(Ontology.IDENTIFIER_KEY) + "#desc");
         Bundle descBundle = Bundle.of(EntityClass.REPOSITORY_DESCRIPTION, descmap);
 
         // Add dates and descriptions to the bundle since they're @Dependent
         // relations.
-        for (Map<String, Object> dpb : extractDates(itemData)) {
+        for (Map<String, Object> dpb : Helpers.extractDates(itemData)) {
             descBundle = descBundle.withRelation(Ontology.ENTITY_HAS_DATE, Bundle.of(EntityClass.DATE_PERIOD, dpb));
         }
 
         //add the address to the description bundle
-        Map<String, Object> address = extractAddress(itemData);
+        Map<String, Object> address = Helpers.extractAddress(itemData);
         if (!address.isEmpty()) {
             descBundle = descBundle.withRelation(Ontology.ENTITY_HAS_ADDRESS, Bundle.of(EntityClass.ADDRESS, address));
         }
-        Map<String, Object> unknowns = extractUnknownProperties(itemData);
+        Map<String, Object> unknowns = Helpers.extractUnknownProperties(itemData);
         if (!unknowns.isEmpty()) {
             logger.debug("Unknown Properties found");
             descBundle = descBundle.withRelation(Ontology.HAS_UNKNOWN_PROPERTY, Bundle.of(EntityClass.UNKNOWN_PROPERTY, unknowns));
         }
-        for (Map<String, Object> dpb : extractMaintenanceEvent(itemData)) {
+        for (Map<String, Object> dpb : Helpers.extractMaintenanceEvent(itemData)) {
             logger.debug("maintenance event found");
             //dates in maintenanceEvents are no DatePeriods, they are not something to search on
             descBundle = descBundle.withRelation(Ontology.HAS_MAINTENANCE_EVENT, Bundle.of(EntityClass.MAINTENANCE_EVENT, dpb));
