+++ ehri-io/src/main/java/eu/ehri/project/importers/xml/DcEuropeanaHandler.java
@@ -66,7 +66,7 @@ private final ImmutableMap<String, Class<? extends Entity>> possibleSubnodes
                     extractIdentifier(currentMap);
                     extractName(currentMap);
                     
-                    Helpers.putPropertyInGraph(currentMap, "sourceFileId", currentMap.get(OBJECT_IDENTIFIER).toString());
+                    Helpers.putPropertyInGraph(currentMap, "sourceFileId", currentMap.get(Helpers.OBJECT_IDENTIFIER).toString());
                     importer.importItem(currentMap, new Stack<>());
 //                importer.importTopLevelExtraNodes(topLevel, current);
                     //importer.importItem(currentGraphPath.pop(), Lists.<String>newArrayList());
@@ -130,11 +130,11 @@ private final ImmutableMap<String, Class<? extends Entity>> possibleSubnodes
     }
 
     private void extractIdentifier(Map<String, Object> currentMap) {
-        if (currentMap.containsKey(OBJECT_IDENTIFIER)){
-        logger.debug(currentMap.get(OBJECT_IDENTIFIER)+"");
-            String id = currentMap.get(OBJECT_IDENTIFIER).toString();
+        if (currentMap.containsKey(Helpers.OBJECT_IDENTIFIER)){
+        logger.debug(currentMap.get(Helpers.OBJECT_IDENTIFIER)+"");
+            String id = currentMap.get(Helpers.OBJECT_IDENTIFIER).toString();
             if(id.startsWith("http://www.beeldbankwo2.nl/detail_no.jsp?action=detail&imid=")){
-                currentMap.put(OBJECT_IDENTIFIER, id.substring(60));
+                currentMap.put(Helpers.OBJECT_IDENTIFIER, id.substring(60));
             }
         }else{
             for(String key: currentMap.keySet())
